+++ src/Private/Mutations.ps1
@@ -15,6 +15,7 @@ function _buildMutation {
         Site { $fieldName = 'site' }
         ApplicationLogin { $fieldName = 'application_login' }
         LoginCredential { $fieldName = 'login_credential' }
+        RecordedLogin { $fieldName = 'recorded_login' }
         EmailRecipient { $fieldName = 'email_recipient' }
         Scope { $fieldName = 'scope' }
         default {}