+++ unit/tests/public/Get-BurpSuiteSiteTree.tests.ps1
@@ -24,7 +24,7 @@ InModuleScope $env:BHProjectName {
 
         It "should add <FieldName> sub selection field" -TestCases @(
             @{ FieldName = "folders"; Query = "folders { id name parent_id }" }
-            @{ FieldName = "sites"; Query = "sites { id name parent_id scope { included_urls excluded_urls } scan_configurations { id } application_logins { id label username } ephemeral email_recipients { id email } }" }
+            @{ FieldName = "sites"; Query = "sites { id name parent_id scope { included_urls excluded_urls } scan_configurations { id } application_logins { login_credentials { id label username } recorded_logins { id label } } ephemeral email_recipients { id email } }" }
         ) {
             # arrange
             Mock -CommandName _callAPI -MockWith {